+++ src/libtrx/game/console/cmd/play_cutscene.c
@@ -1,0 +1,37 @@
+#include "game/console/cmd/play_cutscene.h"
+
+#include "game/game_flow/common.h"
+#include "game/game_string.h"
+#include "strings.h"
+
+static COMMAND_RESULT M_Entrypoint(const COMMAND_CONTEXT *ctx);
+
+static COMMAND_RESULT M_Entrypoint(const COMMAND_CONTEXT *const ctx)
+{
+    int32_t cutscene_to_load = -1;
+    if (String_IsEmpty(ctx->args)) {
+        return CR_BAD_INVOCATION;
+    } else if (String_ParseInteger(ctx->args, &cutscene_to_load)) {
+        cutscene_to_load--;
+        const GF_LEVEL_TABLE *const level_table =
+            GF_GetLevelTable(GFLT_CUTSCENES);
+        if (cutscene_to_load < 0 || cutscene_to_load >= level_table->count) {
+            Console_Log(GS(OSD_INVALID_CUTSCENE));
+            return CR_FAILURE;
+        }
+        const GF_LEVEL *const level = &level_table->levels[cutscene_to_load];
+        GF_OverrideCommand((GF_COMMAND) {
+            .action = GF_START_CINE,
+            .param = cutscene_to_load,
+        });
+        Console_Log(GS(OSD_PLAY_CUTSCENE), level->num + 1);
+        return CR_SUCCESS;
+    } else {
+        return CR_BAD_INVOCATION;
+    }
+}
+
+CONSOLE_COMMAND g_Console_Cmd_PlayCutscene = {
+    .prefix = "cut(scene)?",
+    .proc = M_Entrypoint,
+};