+++ src/libtrx/include/libtrx/debug.h
@@ -10,8 +10,14 @@
         }                                                                      \
     } while (0)
 
-#define ASSERT_FAIL(x)                                                         \
+#define ASSERT_FAIL()                                                          \
     do {                                                                       \
         LOG_DEBUG("Assertion failed");                                         \
+        __builtin_trap();                                                      \
+    } while (0)
+
+#define ASSERT_FAIL_FMT(fmt, ...)                                              \
+    do {                                                                       \
+        LOG_DEBUG("Assertion failed: " fmt __VA_OPT__(, ) __VA_ARGS__);        \
         __builtin_trap();                                                      \
     } while (0)