+++ docs/tr1/CHANGELOG.md
@@ -10,6 +10,7 @@
 - added a fade-out effect when exiting the pause screen to the inventory
 - added exit fade-out effects (#2348)
 - added optional dynamic lighting for gun flashes and explosions, similar to TR2+ (#2357)
+- added a `/cut` console command for playing cutscenes
 - ⚠️ changed the game data to use a separate strings file for text information, removing it from the game flow file
 - changed the object texture limit from 2048 to unlimited (within game's overall memory cap)
 - changed the sprite texture limit from 512 to unlimited (within game's overall memory cap)